Extract Contracts::Core
