+++ lib/contracts.rb
@@ -9,48 +9,15 @@ require "contracts/engine"
 require "contracts/method_handler"
 require "contracts/validators"
 require "contracts/call_with"
+require "contracts/core"
 
 module Contracts
   def self.included(base)
-    common(base)
+    base.send(:include, Core)
   end
 
   def self.extended(base)
-    common(base)
-  end
-
-  def self.common(base)
-    return if base.respond_to?(:Contract)
-
-    base.extend(MethodDecorators)
-
-    base.instance_eval do
-      def functype(funcname)
-        contracts = Engine.fetch_from(self).decorated_methods_for(:class_methods, funcname)
-        if contracts.nil?
-          "No contract for #{self}.#{funcname}"
-        else
-          "#{funcname} :: #{contracts[0]}"
-        end
-      end
-    end
-
-    base.class_eval do
-      # TODO: deprecate
-      # Required when contracts are included in global scope
-      def Contract(*args)
-        self.class.Contract(*args)
-      end
-
-      def functype(funcname)
-        contracts = Engine.fetch_from(self.class).decorated_methods_for(:instance_methods, funcname)
-        if contracts.nil?
-          "No contract for #{self.class}.#{funcname}"
-        else
-          "#{funcname} :: #{contracts[0]}"
-        end
-      end
-    end
+    base.send(:extend, Core)
   end
 end
